103. Binary Tree Zigzag Level Order Traversal - O(n)
Time Complexity - O(n) - since we are traversing all the elements of the tree
Space Complexity - O(2^h) - since 2^h are the number of nodes that are present at height h. Since it is a variant of level order traversal, the complexuty will be the max number of nodes at any height. If h is height of the tree, then the complexity of the above code will be O(2^h)