563. Binary Tree Tilt
Time COmplexity - O(n)
Space Complexity - O(h)

Idea : Calculate the sum of left and right child and return that to root in order to evaluate tilt.This is done in logic() function. I have used logic function because I have to return tilt, not the sum of child subtrees. However we need the child subtree sum.
So logic() function returns the sum of val of childs+val of itself
I have taken a global variable tilt to store the sum of tilts as we proceed in recursion.