565. Array Nesting
Time Complexity - O(n) - Since each element is traversed only once... I mean initially first element is chosen and traversed till the same element is reached(kind of a loop), then we
see the visited array, if it is false, only then we call logic() function, else we do not call that.
Space Complexity - O(n) since visited array is used. We can make this O(1) if instead of using visited array, if we set the
elements in nums to Integer.MAX_VALUE which are visited. This way we can differentiate between the elements which are already visited
and which are yet to be visited.