238. Product of Array Except Self - O(n)
Time Complexity - O(n) - 2 iterations
Space Complexity - O(1) if output which is returned is not considered as extra space

Idea - If input is [1, 2, 3, 4] then we want [2*3*4, 1*3*4, 1*2*4, 1*2*3].  Output can be constructed in two passes. The first pass, skip the first element [None, 1, 1*2, 2*3] = [1, 1, 2, 6] Each element of output is the product of all the elements before it. Then reverse this result and the nums array and apply the same method. [4, 3, 2, 1] --> [1, 4, 4*3, 4*3*2] --> [1, 4, 12, 24] 
Then combine the arrays: [1, 4, 12, 24] and [6, 2, 1, 1]
	[6, 4*2, 24*1, 1] --> [6, 8, 12, 24]. Then reverse it [24, 12, 8, 6]