39. Combination Sum - O(n^t)
Time Complexity - O(n^t)
Let n be the number of elements in the array
First level in the graph will have 0, Second level in graph will have n nodes.(for each element in the array). Further more for lebel 3, there will be n*n nodes as for each node in level 2, there will be n nodes in level 3. Similarly so on
1+n+n^2+n^3+...+n^t =n^t-1/n-1=O(n^t)
where t is the target. Worst case when array has only one element and that value is 1. Height of this tree/grapg will be equal to target t. Rest of the cases, the depth will be less than t.
Space Complexity - O(n)