79. Word Search - O((m*n)^2)
Time Complexity - O((n*m)^2) - n*m for the two for loops, another m*n for recursive function. COnsider a worst case, when we have to traverse all the elements of the matrix(when the given string is not present)

a a a
a b a
a a a

Say the string is aaaaaaaaa

For m*n times, all the m*n are traversed recursively. Therefore O(m*n*m*n) = O((m*n)^2)
Space Complexity - O(m*n)