526. Beautiful Arrangement - O(2^n)
Time Complexity - O(2^n) - It's like constructing an n-ary tree. The height of the tree will be n. So the total elements in the tree will be 2^n
Space Complexity - O(n^2) - n for the recursive stack(length of stack=length of tree=n), n for vis[] array. In every recursive call, we are creating a new array. So the total space consumed = n*n = n^2