537. Complex Number Multiplication - O(1)
Time Complexity - O(1) Since it is given that the given two strings are of the form a+ib where both a and b belongs to [-100,100]. Hence we can assume the length of the given string as constants. So the split can be considered as a constant operation in our case
Space Complexity - O(1) Since it is given that the given two strings are of the form a+ib where both a and b belongs to [-100,100]. So the size of  a1 and b1 is constant.