700. Search in a Binary Search Tree - O(n)
Time Complexity - O(2^h) = O(n) since it is a binary tree and h here is the height of the tree, n is the number of node in the tree. Since all the nodes in the tree are traversed, time complexity is O(n)
Space Complexity - O(h) = O(n) in case of a skewed tree. The recursive stack will have atmost n elements in the stack in case of a skewed tree