112. Path Sum - O(n)
Time Complexity - O(n) where n is the number of nodes. O(n) complexity as all the nodes are traversed in the worst case for finding the appropriate path
Space Complexity - O(h) = O(n) in case of a skewed tree. At any instant, the recursive stack will contain at most h number of nodes in it where h is the height of the tree which is equal to log(n) in general. However, in the case of a skewed tree, h=n.